% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_bbl.R
\name{geo_bbl}
\alias{geo_bbl}
\alias{geo_bbl_data}
\title{Retrieve Geoclient Response for BBLs as a Dataframe}
\usage{
geo_bbl_data(.data, bbl, id = NULL, key = NULL, rate_limit = TRUE)

geo_bbl(bbl, id = NULL, key = NULL, rate_limit = TRUE)
}
\arguments{
\item{.data}{Dataframe containing columns to be used for other arguments.}

\item{bbl}{Either a vector of BBLs (numeric or character is accepted), or a
bare column name of the bbl field if a dataframe is provided.}

\item{id}{The API app ID provided to you from the NYC Developer Portal
formated in quotes. Defaults to \code{NULL} and your key is accessed from your
\code{.Renviron}.}

\item{key}{The API app key provided to you from the NYC Developer Portal
formated in quotes. Defaults to \code{NULL} and your key is accessed from your
\code{.Renviron}.}

\item{rate_limit}{Whether you would like to limit the rate of API requests in
adherence to Geoclient's Service Usage Guidelines. See \code{?geoclient} for
more information.}
}
\description{
This function takes BBLs (borough-block-lot) and returns the Geoclient
response as a tibble. The BBLs can be provided either in a vector as a named
argument or with a dataframe and column name of the BBL field. The Geoclient
API's app ID and key can either be provided directly as arguments, or you
can first use \code{\link[=geoclient_api_keys]{geoclient_api_keys()}} to add them to your \code{.Renviron} file
so they can be called securely without being stored in your code.
}
\details{
For more details see the Geoclient Documentation's guide to
\href{https://api.cityofnewyork.us/geoclient/v1/doc#section-1.2.2}{making BBL requests},
interpreting the \href{https://api.cityofnewyork.us/geoclient/v1/doc#section-2.2}{Geosupport return codes}, the
\href{https://api.cityofnewyork.us/geoclient/v1/doc#section-3.2}{data returned by \code{geo_bbl}},
and a \href{https://api.cityofnewyork.us/geoclient/v1/doc#section-4.0}{complete data dictionary} for
all possible data elements returned by any geoclient function.
}
\examples{

\dontrun{

geoclient_api_keys("1a2b3c4", "9d8f7b6wh4jfgud67s89jfyw68vj38fh")

geo_bbl(1005430053)

library(dplyr)

df <- tibble(BBL = c("1005430053", "1005107502"))

geo_bbl_data(df, BBL)

bind_cols(df, geo_bbl_data(df, BBL))

mutate(df, condo_bbl = geo_bbl(BBL)[["condominiumBillingBbl"]])

}

}
