% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_bin.R
\name{geo_bin}
\alias{geo_bin}
\alias{geo_bin_data}
\title{Retrieve Geoclient Response for BINs as a Dataframe}
\usage{
geo_bin_data(.data, bin, id = NULL, key = NULL, rate_limit = TRUE)

geo_bin(bin, id = NULL, key = NULL, rate_limit = TRUE)
}
\arguments{
\item{.data}{Dataframe containing columns to be used for other arguments.}

\item{bin}{Either a vector of BINs (numeric or character is accepted), or a
bare column name of the bin field if a dataframe is provided.}

\item{id}{The API app ID provided to you from the NYC Developer Portal
formated in quotes. Defaults to \code{NULL} and your key is accessed from your
\code{.Renviron}.}

\item{key}{The API app key provided to you from the NYC Developer Portal
formated in quotes. Defaults to \code{NULL} and your key is accessed from your
\code{.Renviron}.}

\item{rate_limit}{Whether you would like to limit the rate of API requests in
adherence to Geoclient's Service Usage Guidelines. See \code{?geoclient} for
more information.}
}
\description{
This function takes BINs (Building Identification Number) and returns the
Geoclient response as a tibble. The BINs can be provided either in a vector
as a named argument or with a dataframe and column name of the BIN field. The
Geoclient API's app ID and key can either be provided directly as arguments,
or you can first use \code{\link[=geoclient_api_keys]{geoclient_api_keys()}} to add them to your \code{.Renviron}
file so they can be called securely without being stored in your code.
}
\details{
For more details see the Geoclient Documentation's guide to
\href{https://api.cityofnewyork.us/geoclient/v1/doc#section-1.2.3}{making BIN requests},
interpreting the \href{https://api.cityofnewyork.us/geoclient/v1/doc#section-2.2}{Geosupport return codes}, the
\href{https://api.cityofnewyork.us/geoclient/v1/doc#section-3.3}{data returned by geo_bin},
and a \href{https://api.cityofnewyork.us/geoclient/v1/doc#section-4.0}{complete data dictionary} for
all possible data elements returned by any geoclient function.
}
\examples{

\dontrun{

geoclient_api_keys("1a2b3c4", "9d8f7b6wh4jfgud67s89jfyw68vj38fh")

geo_bin(1015862)

library(dplyr)

df <- tibble(BIN = c("1008760", "1007941"))

geo_bin_data(df, BIN)

bind_cols(df, geo_bin_data(df, BIN))

mutate(df, bid = geo_bin(BIN)[["businessImprovementDistrict"]])

}

}
