% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoclient-package.R
\docType{package}
\name{geoclient}
\alias{geoclient}
\alias{geoclient-package}
\title{An interface to NYC's \emph{Geoclient} REST API.}
\description{
An interface to NYC's \emph{Geoclient} REST API.
}
\author{
Maxwell Austensen
}
