% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoclient_api_key.R
\name{geoclient_api_key}
\alias{geoclient_api_key}
\title{Install Geoclient API Key for Repeated Use}
\usage{
geoclient_api_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{The API key provided to you from the NYC Developer Portal
formatted in quotes.}

\item{overwrite}{If this is set to TRUE, it will overwrite the existing
GEOCLIENT_KEY that you already have in your \code{.Renviron} file. And if \code{key}
is set to \code{NULL} it will completely remove that line.}

\item{install}{if TRUE, will install the key in your \code{.Renviron} file for use
in future sessions.  Defaults to FALSE.}
}
\description{
This function will add your Geoclient app ID and key to your \code{.Renviron} file
so they can be called securely without being stored in your code. After you
have installed your key, it can be called any time by typing
\code{Sys.getenv("GEOCLIENT_KEY")} and can be used in package functions by simply
typing GEOCLIENT_KEY. If you do not have an \code{.Renviron} file, the function
will create on for you. If you already have an \code{.Renviron} file, the function
will append the key to your existing file, while making a backup of
your original file for disaster recovery purposes. You can acquire your
Geoclient API Key by first registering with the \href{https://api-portal.nyc.gov/}{NYC's API Portal} at, then adding a "subscription" to the
\href{https://api-portal.nyc.gov/products/geoclient-user}{Geoclient User} API.
}
\examples{

\dontrun{
# To set the keys for use in this session only:
geoclient_api_key(key = "9d8f7b6wh4jfgud67s89jfyw68vj38fh")

# To save the keys in your R.environ file for use in future sessions:
geoclient_api_key(
  key = "9d8f7b6wh4jfgud67s89jfyw68vj38fh",
  install = TRUE
 )

# If you need to overwrite existing keys:
geoclient_api_key("9d8f7b6wh4jfgud67s89jfyw68vj38fh", overwrite = TRUE, install = TRUE)

# First time, reload your environment so you can use the keys without restarting R.
readRenviron("~/.Renviron")

# You can check them with:
Sys.getenv("GEOCLIENT_KEY")

}

}
