% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoclient_bbl.R
\name{geoclient_bbl}
\alias{geoclient_bbl}
\title{Retrieve dataframe response from Geoclient for BBLs}
\usage{
geoclient_bbl(df = NULL, bbl, id = NULL, key = NULL)
}
\arguments{
\item{df}{Dataframe that contains a column of BBLs. Defaults to \code{NULL} and
\code{bbl} is taked as a vector.}

\item{bbl}{Either a vector of BBLs (numeric or character is accepted), or a
bare column name of the bbl field if a dataframe is provided.}

\item{id}{The API app ID provided to you from the NYC Developer Portal
formated in quotes. Defaults to \code{NULL} and your key is accessed from your
\code{.Renviron}.}

\item{key}{The API app key provided to you from the NYC Developer Portal
formated in quotes. Defaults to \code{NULL} and your key is accessed from your
\code{.Renviron}.}
}
\description{
This function takes BBLs (borough-block-lot) and returns Geoclient response
in a tibble. The BBLs can be provided either in a vector as a named argument
or as a dataframe and column name of the BBL field. The Geoclient API's app
ID and key can either be provided directly as arguments, or you can first use
\code{\link[=geoclient_api_keys]{geoclient_api_keys()}} to add them to your \code{.Renviron} file so they can be
called securely without being stored in your code. You can acquire your
Geoclient app ID and Key by first registering with the
\link{https://developer.cityofnewyork.us/user/register?destination=api}(NYC
Developer Portal) at, then
\link{https://developer.cityofnewyork.us/create/project}(create a new project),
selecting "Geoclient v1" from available APIs.
}
\examples{

\dontrun{

geoclient_api_keys("1a2b3c4", "9d8f7b6wh4jfgud67s89jfyw68vj38fh", install = TRUE)
# First time, reload your environment so you can use the keys without restarting R.
readRenviron("~/.Renviron")

geoclient_bbl(bbl = 1005430053)
geoclient_bbl(bbl = c("1005430053", "1005107502"))

library(dplyr)

df <- tibble(BBL = c("1005430053", "1005107502"))

df \%>\% geoclient_bbl(BBL)
}

}
