% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoclient_bin.R
\name{geoclient_bin}
\alias{geoclient_bin}
\title{Retrieve Dataframe Response from Geoclient for BINs}
\usage{
geoclient_bin(df = NULL, bin, id = NULL, key = NULL, rate_limit = TRUE,
  cap_daily_requests = TRUE)
}
\arguments{
\item{df}{Dataframe that contains a column of BINs. Defaults to \code{NULL} and
\code{bin} is taken as a vector.}

\item{id}{The API app ID provided to you from the NYC Developer Portal
formated in quotes. Defaults to \code{NULL} and the id is accessed from your
\code{.Renviron}.}

\item{key}{The API app key provided to you from the NYC Developer Portal
formated in quotes. Defaults to \code{NULL} and the key is accessed from your
\code{.Renviron}.}

\item{rate_limit}{Wether you would like to limit the rate of API requests in
adherence to Geoclient's Service Usage Guidelines. See \code{?geoclient} for
more information.}

\item{cap_daily_requests}{Wether you would like to cap the daily number of
API requests in adherence to Geoclient's Service Usage Guidelines. See
\code{?geoclient} for more information.}

\item{bbl}{Either a vector of BINs (numeric or character is accepted), or a
bare column name of the bin field if a dataframe is provided.}
}
\description{
This function takes BINs (Building Identification Number) and returns the
Geoclient response as a tibble. The BINs can be provided either in a vector
as a named argument or with a dataframe and column name of the BIN field. The
Geoclient API's app ID and key can either be provided directly as arguments,
or you can first use \code{\link[=geoclient_api_keys]{geoclient_api_keys()}} to add them to your \code{.Renviron}
file so they can be called securely without being stored in your code.
}
\examples{

\dontrun{

geoclient_api_keys("1a2b3c4", "9d8f7b6wh4jfgud67s89jfyw68vj38fh")

geoclient_bin(bin = 1015862)
geoclient_bin(bin = c("1008760", "1007941"))

library(dplyr)

df <- tibble(BIN = c("1008760", "1007941"))

geoclient_bin(df, BIN)
}

}
