% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoclient_search.R
\name{geoclient_search}
\alias{geoclient_search}
\title{Retrieve Dataframe Response from Geoclient for Any Location Type}
\usage{
geoclient_search(df = NULL, location, id = NULL, key = NULL,
  rate_limit = TRUE, cap_daily_requests = TRUE)
}
\arguments{
\item{df}{Dataframe that contains a column of BBLs. Defaults to \code{NULL} and
\code{bbl} is taken as a vector.}

\item{location}{Any of the 6 locations types from other functions: Address,
BBL, BIN, Blockface, Intersection, or Place. The argument can be either a
single vector of locations, or a bare column name of the location field if
a dataframe is provided.}

\item{id}{The API app ID provided to you from the NYC Developer Portal
formated in quotes. Defaults to \code{NULL} and your key is accessed from your
\code{.Renviron}.}

\item{key}{The API app key provided to you from the NYC Developer Portal
formated in quotes. Defaults to \code{NULL} and your key is accessed from your
\code{.Renviron}.}

\item{rate_limit}{Wether you would like to limit the rate of API requests in
adherence to Geoclient's Service Usage Guidelines. See \code{?geoclient} for
more information.}

\item{cap_daily_requests}{Wether you would like to cap the daily number of
API requests in adherence to Geoclient's Service Usage Guidelines. See
\code{?geoclient} for more information.}
}
\description{
This function takes any of the 6 location types as a single input and returns
the Geoclient response as a tibble. The locations are provided either in a
single vector as a named argument or with a dataframe and column name of the
location field. The Geoclient API's app ID and key can either be provided
directly as arguments, or you can first use \code{\link[=geoclient_api_keys]{geoclient_api_keys()}} to add
them to your \code{.Renviron} file so they can be called securely without being
stored in your code.
}
\details{
Geoclient is quite flexible with the format of the address when
provided as a single input for this function, however the results may be
slower than the more restrictive \code{\link[=geoclient_address]{geoclient_address()}} because Geoclient
make have to make multiple requests to narrow in on the correct location.

For more details see the Geoclient Documentation's guide to \href{https://api.cityofnewyork.us/geoclient/v1/doc#section-1.3}{single-inputrequests},
interpreting the \href{https://api.cityofnewyork.us/geoclient/v1/doc#section-2.2}{Geosupport returncodes}, and a
\href{https://api.cityofnewyork.us/geoclient/v1/doc#section-4.0}{complete datadictionary} for
all possible data elemets returned by any geoclient function.
}
\examples{

\dontrun{

geoclient_api_keys("1a2b3c4", "9d8f7b6wh4jfgud67s89jfyw68vj38fh")

geoclient_search(location = 1005430053)
geoclient_search(location = "139 macdougal st mn")
geoclient_search(location = c("1008760", "1007941"))

library(dplyr)

df <- tibble(loc = c("1005430053", "139 MacDougal Street, New York, 10012", "1008760"))

geoclient_search(df, loc)
}

}
