% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoclient-package.R
\docType{package}
\name{geoclient-package}
\alias{geoclient}
\alias{geoclient-package}
\title{geoclient: Tools to work with NYC's _Geoclient_ REST API}
\description{
Tools to work with NYC's _Geoclient_ REST API. Retrieve Borough-Block-Lot (BBL) and other
data from addresses. Allows users to optionally return a simple features (sf) dataframe with a
list-column containing BBL geometry.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/austensen/geoclient}
  \item Report bugs at \url{https://github.com/austensen/geoclient/issues}
}

}
\author{
\strong{Maintainer}: Maxwell Austensen \email{maxwell.austensen@nyu.edu}

}
\keyword{internal}
